+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-files: ^(metricflow|tests)/
+files: ^(metricflow|tests|dbt\-metricflow\/dbt_metricflow)/
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.2.0
