+++ dbt-metricflow/pyproject.toml
@@ -24,12 +24,25 @@ classifiers = [
   "Programming Language :: Python :: Implementation :: PyPy",
 ]
 dependencies = [
+  # cli dependencies
+  "Jinja2>=3.1.3",
+  "click>=7.1.2",
+  "halo>=0.0.31, <0.1.0",
+  "update-checker>=0.18.0, <0.19.0",
+
+  # Internal dependencies
   "dbt-core>=1.7.4, <1.8.0",
-  "metricflow>=0.205.0, <0.206.0"
+  "metricflow>=0.205.0, <0.206.0",
+
+  # dsi version should be fixed by MetricFlow/dbt-core, not set here
+  "dbt-semantic-interfaces",
 ]
 
 [project.urls]
 "Source Code" = "https://github.com/dbt-labs/metricflow/tree/main/dbt-metricflow"
+
+[project.scripts]
+mf = 'dbt_metricflow.cli.main:cli'
 
 [project.optional-dependencies]
 bigquery = [
