+++ tests/cli/test_cli.py
@@ -15,8 +15,8 @@ from dbt_semantic_interfaces.parsing.dir_to_model import (
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 from dbt_semantic_interfaces.test_utils import base_semantic_manifest_file
 
-from metricflow.cli.cli_context import CLIContext
-from metricflow.cli.main import (
+from dbt_metricflow.cli.cli_context import CLIContext
+from dbt_metricflow.cli.main import (
     dimension_values,
     dimensions,
     entities,
